pieces placement

+++ proj/src/game/sprite/sprite.c
@@ -103,7 +103,7 @@ int draw_text_sprite(sprite_t *sprite, int color_map) {
   return EXIT_SUCCESS;
 }
 
-int draw_piece_in_mode_14c(uint8_t *map, uint8_t x, uint8_t y, unsigned int size) {
+int draw_piece_in_mode_14c(uint8_t *map, int x, int y, unsigned int size) {
 
   for (unsigned int height = 0; height < size; height++) {
     if (height + y >= get_vres())
@@ -113,6 +113,7 @@ int draw_piece_in_mode_14c(uint8_t *map, uint8_t x, uint8_t y, unsigned int size
       if (width + x >= get_hres())
         break;
 
+      // TODO::: WE CAN MAKE THIS HARDCODED AND SAVE SOME TIME
       RGB rgb = RGB_new(0);
 
       rgb.setBlue(&rgb, *map);
